fixed to runnable state

+++ src/main/java/me/jweissen/aeticket/dto/response/UserResponseDto.java
@@ -1,4 +1,4 @@
 package me.jweissen.aeticket.dto.response;
 
-public record UserResponseDto(int id, String firstname, String lastname, String email) {
+public record UserResponseDto(Long id, String firstname, String lastname, String email) {
 }
